refactor: GridItem 분리 #43
